requirements.txt
bitalino==1.2.6
bleak==1.1.0
numpy==2.3.2
pandas==2.3.2
PyBluez-bitalino==0.30
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.1
six==1.17.0
tomli==2.0.1
typing_extensions==4.15.0
tzdata==2025.2
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Advertisement==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Storage.Streams==3.2.1
